Move test util to util package
